Resolved TODO move addLegend from map.ts to print.service.ts Fix reverse printDimension (bad printing format A0 to A5) Improve legend size
